Address size_t/uint64_t conversion warnings
Address conversion from uint64_t to size_t possible loss of data
warnings for 32-bit builds.
- Add a macro to assert that the uint64_t value is less than or equal to
  the max size_t size.
- Add data loss macro and static_cast for uint64_t to size_t conversion.

+++ util/page_guard_manager.cpp
@@ -200,7 +200,7 @@ size_t PageGuardManager::GetSystemPageSize() const
 
 size_t PageGuardManager::GetAdjustedSize(size_t size) const
 {
-    uint64_t extra = size % system_page_size_;
+    size_t extra = size % system_page_size_;
     if (extra != 0)
     {
         // Adjust the size to be a multiple of the system page size.
